pyproject.toml
[project]
name = "ubidtours-pipeline"
version = "0.1.0"
description = "A pipeline to sync data from MySQL to Pinecone."
requires-python = ">=3.8"
dependencies = [
    "sqlalchemy",
    "mysql-connector-python",
    "sentence-transformers",
    "pinecone",
    "python-dotenv",
    "numpy",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = [
  "."
]
